+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.0.2
+
+* [Feat] Add method `LanguageHelper.instance.analyze()` to analyze the missing texts of all languages.
+* Add headers for `LanguageCodes` enum.
+
+
 ## 0.0.1+1
 
 * [LanguageNotifier] Only change the state of the root widgets to improve performance.